Refactor out apply_for_court_date classification
