+++ lib/hackney/income/tenancy_classification/v2/classifier.rb
@@ -25,12 +25,11 @@ module Hackney
               Rulesets::SendInformalAgreementBreachLetter,
               Rulesets::InformalBreachedAfterLetter,
               Rulesets::SendCourtAgreementBreachLetter, # TODO(AO): Possible missing test for this classification
-              Rulesets::SendCourtWarningLetter
+              Rulesets::SendCourtWarningLetter,
+              Rulesets::ApplyForCourtDate
             ]
 
             actions = rulesets.map { |ruleset| ruleset.new(@case_priority, @criteria, @documents).execute }
-
-            actions << :apply_for_court_date if apply_for_court_date?
 
             actions.compact!
 
@@ -62,24 +61,6 @@ module Hackney
             raise ArgumentError, "Tried to classify a case as #{wanted_action}, but this is not on the list of valid classifications."
           end
 
-          def apply_for_court_date?
-            return false if should_prevent_action?
-            return false if @criteria.balance.blank?
-            return false if @criteria.weekly_gross_rent.blank?
-            return false if @criteria.active_agreement?
-
-            return false unless @criteria.nosp.served?
-
-            return false unless @criteria.last_communication_action.in?(valid_actions_for_apply_for_court_date_to_progress)
-            return false if last_communication_newer_than?(2.weeks.ago)
-
-            return false unless @criteria.nosp.active?
-
-            return false if @criteria.courtdate.present? && @criteria.courtdate > @criteria.last_communication_date
-
-            balance_is_in_arrears_by_number_of_weeks?(4)
-          end
-
           def case_has_eviction_date?
             @criteria.eviction_date.present?
           end
@@ -90,12 +71,6 @@ module Hackney
 
           def case_paused?
             @case_priority.paused?
-          end
-
-          def valid_actions_for_apply_for_court_date_to_progress
-            [
-              Hackney::Tenancy::ActionCodes::COURT_WARNING_LETTER_SENT
-            ]
           end
 
           def active_agreement_court_outcomes
